requirements.txt
langchain
langchain[openai]
langchain[anthropic]
langchain-core
langgraph
langchain-community
tavily-python
langgraph-checkpoint-sqlite
python-dotenv
beautifulsoup4
langchain-text-splitters

